Removed unused import, renamed dictionary data folder/file
