+++ src/dictionary/dictionary.ts
