+++ src/store/api/apiSlice.ts
@@ -1,5 +1,4 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import { Movie } from './apiTypes';
 
 const BASE_URL = 'http://localhost:4000';
